+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Generalized support for setting (default) OIDC provider id through env var `OPENEO_AUTH_PROVIDER_ID`
   [#419](https://github.com/Open-EO/openeo-python-client/issues/419)
+- Added `OidcDeviceCodePollTimeout`: specific exception for OIDC device code flow poll timeouts
 
 ### Changed